Updates to formating, moved export-porfile file
