+++ src/app/feature/consultant/components/consultant-header/consultant-header.component.html
@@ -2,7 +2,7 @@
   <ng-container *ngIf="consultant$ | async as consultant">
     <div class="picture-wrapper">
       <div class="button-container">
-        <button mat-fab class="export-button" style='background-color: green' (click)="exportProfile(consultant)">
+        <button mat-fab class="export-button" (click)="exportProfile(consultant)">
           <mat-icon>arrow_downward</mat-icon>
         </button>
         <button mat-fab class="edit-button" color="white" (click)="openEditHeaderDialog()">
